+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 5.1.0 (unreleased version)
+#### Notes
+Extends support of the SDK to OneView REST API version 800, 1000 and 1200.
+
+#### Features supported
+- Hypervisor Managers
+
 # 5.0.0
 #### Notes
 Extends support of the SDK to OneView REST API version 1200 (OneView v5.00).